Modified backend so that the app can call external APIs for easier media searching

+++ backend/backend/src/main/java/com/dypaworld/service/UserService.java
@@ -11,7 +11,7 @@ public interface UserService {
 
     User updateUserDetails(UserDTO userDTO);
 
-    boolean deleteUser(Integer userId);
+    boolean deleteUser(Long userId);
 
-    User getUserById(Integer userId);
+    User getUserById(Long userId);
 }
